# src/swaggerDocument.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "DLink cameras manager API",
    "description": "Документация по апи камер DLink",
    "version": "0.0.1"
  },
  "paths": {
    "/api/v1/status": {
      "get": {
        "summary": "Статус сервиса",
        "description": "",
        "responses": {
          "200": {
            "description": "Статус состояния сервиса",
            "content": {
              "application/json": {
                "example": {
                  "status": "active"
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cameras/{cameraId}": {
      "patch": {
        "summary": "Управление сенсорами камеры",
        "description": "Ручка для управления датчиками, доступными камере. Для правильной работы нужна инициализация камеры в настройках",
        "parameters": [
          {
            "in": "path",
            "name": "cameraId",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "ID камеры"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "event": "enableSoundSensor"
              },
              "schema": {
                "type": "object",
                "properties": {
                  "event": {
                    "type": "string",
                    "enum": [
                      "enableSoundSensor",
                      "disableSoundSensor",
                      "disableMotionSensor",
                      "enableMotionSensor"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Состояние успешно изменено",
            "content": {
              "application/json": {
                "example": {},
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка вводных данных",
            "content": {
              "application/json": {
                "example": {
                  "error": "Сообщение ошибки валидации"
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Ошибка при попытке найти данные",
            "content": {
              "application/json": {
                "example": {
                  "error": "Текст ошибки"
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cameras": {
      "get": {
        "summary": "Получить все настройки камер",
        "responses": {
          "200": {
            "description": "Состояние успешно изменено",
            "content": {
              "application/json": {
                "example": {
                  "data": [
                    {
                      "id": "000000235900",
                      "isSoundSensorEnabled": false,
                      "isMotionSensorEnabled": true
                    },
                    {
                      "id": "000000235900",
                      "isSoundSensorEnabled": true,
                      "isMotionSensorEnabled": true
                    }
                  ]
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "isSoundSensorEnabled": {
                            "type": "boolean"
                          },
                          "isMotionSensorEnabled": {
                            "type": "boolean"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Камеры не настроены",
            "content": {
              "application/json": {
                "example": {
                  "error": "Камеры не настроены"
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}